Make UDP sockets more POSIX conform.
We now have a connect operation that sets a default destination for
sends and implicitly bind the socket to an ephemeral port on bind
(0), connect, and send_to. Due to the default destination we now also
offer a send operation and, for consistency, a recv operation that does
not deliver the remote endpoint to the caller.