tilemux: introduce an ActivityRef type that tracks refs at runtime.
Unfortunately, it's not really possible to use a StaticRefCell or
similar for all activities, because we sometimes require access to two
activities at a time. Therefore, we now use an unsafe cell, but track
the references per activity at runtime to ensure that there are never
two mutable references to the same activity at the same time.