host: added a new build mode 'coverage' for code coverage.
Supporting code coverage is rather easy for C++, because we just need to
build everything with --coverage. It's a bit more tricky for Rust,
because it normally depends on a compiler-provided crate
(profiler_builtins), but for some reason we cannot build it. However,
we can specify that we don't want to use any crate for that and instead
make sure that the required functions that llvm generates calls to are
coming from elsewhere. For that we build our own version of the profile
part of compiler-rt from llvm. This is used by profiler_builtins as
well, so we basically just do the same.

Note that we use our own repository for the profile part of compiler-rt
(in turn a part of llvm), integrated as a submodule here, to avoid that
we need to pull in the entire llvm repository. For that reason, we need
to update these files on every update our the used Rust toolchain.

Note also that we use a separate Rust target for the coverage mode. On
the one hand, Rust only supports release and debug, and we simply use
debug in coverage mode. On the other hand, we want to generate
different files, so that we need to distinguish these two. The best way
seems to be the introduction of another target (equivalent except for
the name).