net: accept packets with possibly wrong checksums (on gem5).
We shouldn't deny packets just because we don't know whether the
checksum is valid. Since we only rely on the e1000 NIC to tell us
whether the checksum is valid and it does not support all protocols, we
should just accept packets where the NIC doesn't know. One example are
ICMP packets. We could check those in software, but it's not important
for now.