Added support for larger sends than our event channel capacity.
Previously, we were assuming that all calls to{TCP,UDP}Socket::send were
passing data that is smaller than the maximum message size of the event
channel. That's okay for UDP and raw sockets, but should of course
yield an error, which it does now. For TCP, we need to split the send
into multiple packets, which is happening now. I also added tests for
all of these cases.