librustm3: restructured activity abstractions.
Instead of having one Activity type that is used to represent the own
and child activities, we now distinguish between the two. The problem
was that many methods only worked properly for either of these cases,
which causes confusion. Thus, OwnActivity now represents the own
activity and only offers the methods that actually make sense for the
own activity. Similarly, ChildActivity represents a child activity.

For example, we do no longer have a FileTable in the ChildActivity,
because these files cannot be used anyway. They old serve as a
temporary storage to know which files to delegate to the child on
run/exec. For that reason, we know simply keep a table of file
descriptors to know what to delegate. This has also the positive side
effect that we do no longer need to store files via Rc<RefCell<dyn
File>>, but can simple story the actual file in a single place via
Box<dyn File>. This in turn simplifies other things like giving out
normal file references to applications.