librustm3: use LazyReadOnlyCell for Activity::own.
I think I have finally found a good trade-off between safety and
usability for the own activity and most importantly files. The root of
the problem is that we have stdin/stdout/stderr and want println to
work everywhere and without needing anything (i.e., with expecting a
reference to stdout as an argument). This in turn requires stdout to be
global, which in turn requires all files and the filetable to be
global. If the filetable is global though, we need to put it into a
RefCell to get mutable access to files or add/remove files (which we
need).

At the same time, using files should be both convenient and robust. It
should be convenient in the sense that applications shouldn't need to
call borrow() every time they want to do something with a file. It
should be robust in the sense that we don't have undefined behavior and
that we don't easily run into runtime panics. These requirements rule
out a couple of other options such as our previous solution
(undefined behavior due to the UnsafeStaticCell for Activity::own) or a
filetable that simple hands out a RefCell<File> (inconvenient).

I also considered to keep a reference/pointer in FileRef and tracking
references on a per-file basis manually, because that would allow us to
hand out normal references while FileRef exists. However, this approach
would mean that FileRef blocks any other access to the file (or possibly
even the entire filetable). This does not work due to println, which
would block access to stdout during the entire runtime. Besides this
problem, I think it would be rather fragile, because it would be easy
to accidentally get access to the same file and thereby run into a
runtime panic.

Instead, the FileRef type holds a file descriptor and can give access to
a file on demand. Therefore, holding a FileRef does not block the
access to the file or the filetable. However, for that reason actually
accessing the file requires us to request it from the filetable and we
cannot hand out normal references, because FileRef does not block the
access to the file. This implies that we cannot implement
Deref/DerefMut for FileRef. Instead, we implement all relevant traits
for files and call borrow() in the implementation ourself. For sockets,
I introduced the traits StreamSocket and DGramSocket for that purpose.
For that reason, using files is convenient, because you can simply call
file.read(), but also safe due to the borrow/RefCell-check behind the
scenes.

Note also that, as before, FileRef has a generic parameter that is used
to cast it to the actual type of the file it provides access to. This
is done via as_any, which all files implement. This trick allows us to
let TcpSocket::new(), VFS::open, etc. create a FileRef with the proper
generic parameter, so that we can directly use all methods of the
concrete file type. If you use FileTable::get() though, you get a
generic FileRef, which only provides access to all methods of File.