Added DNS support.
We have now a DNS type on the client side that can be used to resolve
host names to IP addresses. The net service provides multiple session
arguments to support DNS. At first, we have to specify the nameserver
via "-n <IP>". Second, to reach hosts outside of our local network, we
can specify a gateway via "-g <IP>". And finally, we can define the
network mask via "-a <mask>".

So far, the tools ping and netcat make use of DNS.