net: poll the NIC again if we received multiple packets.
This is necessary for the e1000 driver, because after receiving one
packet and triggering an interrupt, the NIC might receive another
packet before the interrupt got ACKed. In this case, the NIC will not
raise an interrupt again and it is expected that the driver walks
through the descriptor ring.