resmng: fixed bug in case no child keeps the domain tile object.
Previously, if no child was directly using the tile object for the
entire domain (potentially shared among multiple childs), the resource
manager accidentally revoked it. This can happen if we for example have
just one child and the child has specific time/EP/PT requirements and
thus needs a derived tile object.

We therefore now keep the domain tile object (an Rc) to ensure that we
don't revoke it until all childs using it are gone.