Added support for the Rust standard library.
Based on musl and our C-based compatibility layer between the
musl-m3-backend and libm3, we can now support the Rust standard
library. We do so by pretending that we build for Linux and then use
musl at runtime to translate Linux syscalls to the corresponding
operation on M³.

There are a few other tweaks required to support std:

1. We now have the crates m3impl and m3 with the difference that m3impl
does *not* depend on heap and the language items. For that reason, we
can now use std, which also provides the heap (__rdl_alloc etc.) and
the language items (panic handler etc.). The language items are also
provided by the lang crate, which is used by the m3 crate and
applications that don't use the m3 crate.

2. We needed to adjust our target specs to build std correctly. This
requires to pretend that the OS is Linux, our environment is musl, and
we belong to the unix family. Additionally, "linux" needs to be part of
the target triple, because the build.rs of std checks that to determine
whether "restricted-std" is set during build or not (disables network
support etc.). For RISC-V and ARM, we also needed to make sure that we
build with soft-float to prevent that the linker complains about mixes
of hard-float and soft-float.

3. We needed to update to a newer Rust version, because the old one
(2022-04-03) had still bugs in the RISC-V support in std. With the new
one, everything seems to be fine.