build: fixed rebuild/relink problem of Rust programs.
We now use the restat option of ninja rules instead of touching the
output files after running the command. The new cargo version
apparently rebuilds everything if the output file changed. For that
reason, we rebuilt everything as soon as anything changed.

The fix and also the much better solution than before is to use the
restat option of ninja. This option lets ninja re-stat the commands
output after running it any does not rebuild children in the dependency
tree for the outputs that didn't change. Therefore, it's better than
before where due to the touch of all output files we simply assumed
that all output files changed and therefore relinked everything.