rust: added code-coverage support on gem5.
This is based on the "-Cinstrument-coverage" flag and on the minicov
crate. The flag tells Rust to instrument the code accordingly and
minicov is used to extract the information afterwards, which internally
uses llvmprofile.

To support it, we also needed to rename the RISC-V target triples
from "riscv-*" to "riscv64-*", because otherwise the clang invocation
inside of minicov fails as the target "riscv-*" is not known. I think
what's going on is that it checks for certain parts of the target
triple and tries to figure out the llvm target triple based on that and
if that fails, it passes our target triple unchanged to clang (which of
course fails). By renaming it to "riscv64-*", the translation magic
apparently works.

Note that this is only supported for Rust and for RISC-V and x86-64.
Unfortunately, it doesn't work for TileMux, because the resulting
profraw file is always corrupted afterwards.