Removed support for M3_TARGET=host.
Due to its small benefit and the number of limitations that the host
target nowadays has, I've decided to remove it entirely.

In more detail, the primary reasons for the host targets were:
1. better debugging support (valgrind, sanitizers, code coverage, ...)
2. faster execution than on gem5

However, valgrind and sanitizers mostly help for unsafe languages. Now
that most things are written in Rust, they provide little value. Code
coverage is nice, but helps only to some degree on host due to its
limitations (see below) and can very like be supported on gem5 (or even
hw) as well. Faster execution used to be an argument, but now that we
have the hw target, we can just use the FPGA.

On the other side, the host target has several limitations. Most
features that have been added to M³ in recent years such as virtual
memory, context switching, and accelerators are not support on the host
and I don't see a reasonable way to support it. Therefore, many
applications and tests simply don't work (reducing the value of code
coverage, for example). Additionally, anything that relies on the C
library (musl) does not work on host: bsdutils, leveldb, flac, the Rust
standard library, etc. Maybe it's possible to support that on host, but
I don't see a reasonable way.

Due to these limitations and the vast differences between kachel and the
host, the host also becomes increasingly an obstacle when trying to
make progress. Whenever we want to introduce something that relies on
the C library, for example, we either have need some hack to make it
work on host or just need to live with the fact that it doesn't work
there. For example, I'm considering to use libfmt for formatted output
in C++, but that relies on the C library. For something basic like
output, we simply cannot ignore the host target. So either we find some
hack to use it there, or we just don't use it.

There are still small advantages left like that sanitizers are still
helpful for the C/C++ parts of M³ or that another platform helps to
find bugs. But considering the limitations, it seems that it's not
worth the maintenance effort anymore. Therefore, I've now decided to
remove it. That does not only decrease the maintenance effort, but also
vastly simplifies the code in some places and removes lots of dark
corners.