libm3: added type-safe and compile-time checked print.
The basic idea is taken from libfmt, but significantly simplified,
completely based on C++20, and most notably, not relying on the C++
standard library. The last point is important for us, because we want
to use it in baremetal applications as well and it should have as
little code footprint as possible.

The new formatting API provides a type-safe way to formatted printing,
where the format string is parsed and checked at compile time. This
combines the advantages of the classical printf (simple and readable,
suitable for i18n) with the advantages of C++ iostreams (type safe and
less error prone). The API is basically the same as in Rust and other
modern languages: the format string can contain '{}' to refer to
arguments in the given order or '{n}' to refer to the argument with
index n. Additionally, a ':' and further specifications can precede
the '}' to define the way the argument is formatted, such as '{:#x}'.
So, a complete example looks like the following:

println("my arguments {1:.3} and {0:#010x}"_cf, 0x1234, "foobar");

Note that the _cf suffix for the format string is mandatory. This is a
user-defined literal, which converts the string into an string object
(m3::detail::CompiledString) that can be parsed at compile time.

Formatting of custom types is also supported via either adding a
non-static member function called format (with appropriate arguments)
to the type or by adding a template specialization for
Formatter<YourType>.

Besides println, we currently support various other variants for the
different use cases we have: print, print_to, format, format_to, eprint,
eprintln, exitmsg, panic, and vthrow.

With this approach and in contrast to printf, the types of the arguments
are determined by the arguments itself instead of by the format string,
making it type safe. Additionally, the format string is parsed and
therefore syntax checked (e.g., you cannot accidentally forget a '}',
specify format flags in the wrong order or use unsupported flags). At
the same time, the API is as simple as printf and supports i18n.

According to my tests, this approach has less runtime costs and leads to
almost exactly the same code size as before. The compile time for a
complete rebuild is about 25% longer. But I think this is acceptable
given the advantages and the frequency of complete rebuilds.