gem5: updated to v22.0.0.1.
Unfortunately, this breaks ARMv7 for us. gem5 seems to be phasing out
ARMv7 support in favor of ARMv8 support and apparently there are some
issues now with ARMv7 (not sure if they are easy to fix). However, we
should move to ARMv8 at some point as well, I think. But for now we
leave it in a broken state.