libm3: removed buffer-size arguments from RecvGate::new_bind.
Requiring these arguments for bind meant that we needed to specify them
both at creation and for bind. It was therefore easy to get wrong and
also was in contrast to the bind functions for other gate types.

I've now introduced a system call to retrieve the buffer size from an
existing RecvGate, which is used internally in RecvGate to fetch it
once and cache it locally. Unfortunately, we cannot fetch it on
construction, because we do not always have the capability when
creating the RecvGate object for a specific selector. We therefore fetch
it lazily instead.