libm3: auto-activate RecvGates.
So far, only SendGates and MemGates were auto-activated, but RecvGates
needed to be activated manually. This was confusing and actually not
necessary. Since the kernel blocks the activation of SendGates in case
the corresponding RecvGate is not activated yet anyway, this problem
does not need to be handled by applications. For that reason, RecvGates
are now auto-activated on first use like the other gates.

There are still cases where a manual activation is required though. For
example, when we want to configure the send EP for a device that
connects to a receive EP at the driver's tile we need to ensure that
the receive EP exists before configuring the send EP. Similarly, the M³
kernel expects the RecvGate of servers to be activated before
registering the service. But these corner cases are all handled within
libm3 and applications don't have to worry about that.