Use Result/Errors::Code for program/closure results.
That is, Rust main functions and closures now return Result<(), Error>
instead of i32. This allows to use the ? operator in these functions
and also makes it cleaner, because we don't work with integers anymore,
but with actual error codes. C++ main functions still need to return
int, but can also return Errors::Code, because it implicitly converts
to int. For cases that don't have a specific error code, I've
introduced Code::Unspecified. For example, our musl translator uses
Code::Unspecified for exit code 1 and Code::None for exit code 0.