Added synchronous uni-direction channel abstraction.
This is based on the work of Sebastian Ertel with some improvements and
simplifications. The basic idea is to provide an API similar to
std::sync::mpsc::sync_channel, which is simpler to use than a manual
usage of SendGates and RecvGates, but less flexible (always synchronous,
happens between exactly one sender and one receiver, and only one data
type at a time).

A channel consists of a sender and a receiver and allows the sender to
send messages to the receiver. These messages are not buffered, but
delivered synchronously to the receiver. Therefore, our channel is
comparable to a std::sync::mpsc::sync_channel with a bound of 0.