resmng: use separate MsgBuf and structs for messages.
We cannot use the default MsgBuf when communicating with the resource
manager. This is because activating the RecvGate and thereby allocating
a receive buffer requires communication with the resource manager.
Thus, sending a message with the default MsgBuf and with a
non-activated RecvGate for replies could again require the default
MsgBuf to allocate the receive buffer for the reply gate from the
resource manager. For that reason, I've introduced a new MsgBuf
specifically for the communication with the resource manager.

While doing that I also introduced structs for the messages that are
exchanged with the resource manager, because that's the arguably
cleaner way and also prevents us from needing yet another version of
the send_* macros that use a custom MsgBuf.