git: make all submodule paths absolute.
Unfortunately, there does not seem to be a perfect solution, but this is
the most reasonable one apparently. Relative paths - which we had
recently - have the advantage that we can easily move between
repositories and everybody can freely choose between SSH and HTTPS.
However, it also caused trouble multiple times, where people then have
to manually edit the .git/config to fix the URL. It also makes it hard
to fork M³, because you need to fork all submodules as well to make the
paths work.

Absolute paths on the other hand have the advantage to cause less
trouble and allow to easily fork M³, but the disadvantage that people
cannot really choose between SSH and HTTPS anymore. This seems to be
still possible with url.insteadOf in the git config though. Therefore,
I think the most reasonable solution is to use absolute URLs with SSH
and either require people to give an SSH key to Github or use
url.insteadOf.