Added support for multi-chip configurations.
This is based on 14-bit ids that consists of both a chip and a tile id.
We represent these ids with the TileId type to keep the exact encoding
at one place. However, it is currently only supported with gem5.

We also have one example configuration (config/multichip.py) that
creates two chips connected via bridges. Therefore, we can also
simulate the increased latency for cross-chip communication.