Pass FS images as boot modules.
Previously, FS images were always stored at the beginning of the first
memory tile and used completely separate and primitive mechanisms than
boot modules. For example, we used the physmem element in boot scripts
to gain access to FS images, but this had multiple problems. At first,
we needed to know the address and size of the FS images, which is why
we had special variables ($fs.size, ...) in boot scripts. Second, we
couldn't specify the tile it's stored in. Furthermore, the address ways
always hard-coded to 0 at multiple places. Multiple FS images were
possible, but they had to be the same.

Therefore I've now reworked this whole concept and decided to use boot
modules to pass FS images to the kernel. This has several advantages
actually. At first, we can reuse the module infrastructure and don't
need yet another mechanism. Second, boot modules have names, which
allows us to refer to them by name later (instead of needing to know
addresses and sizes). Third, we can easily have multiple ones, both
duplicates and different images. Fourth, we have less reserved areas
with unknown sizes. And fifth, we solve the issue that FS images need
to be in the first memory module.

To make that work, I extended config files to contain a "mods" section
that define additional boot modules to pass to the kernel. For now,
these are always FS images, but of course any file is supported.
Applications need to request access to these modules via
<mod name="<modname>" />.
Optionally, permissions can be reduced.