boot: allow overwriting env vars that are set in boot scripts.
Previously, we had it in the other way around: boot scripts overwrite
the variables set by the user and we warned if that happened. But for
tests and benchmarks we need the functionality to overwrite the
variables in boot scripts.

Therefore, we now don't let scripts overwrite variables but instead warn
if the script tried to set a variable to a different value.