resmng: replace FnMut argument for Subsystem::start with trait.
Having a trait gives us more flexibility to let users of the resmng
crate customize the start of child activities. We currently use that to
additionally allow root and pager to configure additional PMP EPs for a
specific domain.

That solution is much nicer than the previous one where the pager
installed these EPs on a per-activity basis and needed to make sure by
itself to do that only once per tile. Worse, for root we configured the
PMP EP for the boot modules within the resmng crate instead of in root
and we gave (read-write) access to all boot modules.

The new solution avoids the tracking in the pager and allows us to move
the PMP-EP creation for boot modules to root. Additionally, root now
only provides access to the minimum amount of boot modules (using a
single PMP EP). That is, without tile sharing the tile gets only access
to the application's boot module and with tile sharing the tile gets
access to the range between the first and last boot module that is run
on the tile. Fortunately, these are always contiguous at the moment due
to the way we determine the list of boot modules.