libm3: make abort, exit, wait, and sleep static members of OwnActivity.
In C++, these were previously static members of Activity and in Rust
these were non-static members of OwnActivity. The latter also had the
downside that it required the OwnActivity instance, which might not yet
exist. Thus, errors that occurred early lead to a panic in abort,
because abort depended on the OwnActivity instance.