libm3: introduced "compat" description for Tile::get.
Most of the time, we don't care about small differences (BOOM vs.
Rocket, tile with NIC or without, etc.) and only want to find another
compatible tile. Therefore, in contrast to "clone" which searches for
an identical tile, "compat" searches for a compatible tile that has the
same tile type and ISA.