kernel: disallow PMP changes if activities exist.
This is motivated by the problem that we cannot sensibly handle PMP
failures. Since PMP accesses are caused by last-level cache misses, the
code that actually caused the access was executed potentially long ago.
The TCU has therefore no information about the cause of the PMP access
and can therefore not provide any information to the software.

For that reason, we cannot kill the application that triggered the PMP
failure or similar, because we don't know which application actually
caused it. In consequence, the best option seems to be to disallow
changes to PMP EPs after activities have been started *and* (what we
already do) disallow virtual-memory mappings to physical memory that
cannot be accessed.