b: added clean and distclean commands.
These are not required under normal circumstances, because the build
system tracks all dependencies so that a build should always be
correct. However, there are still corner cases where this might be
desirable to have. For example, if your host compiler or libraries were
updated.

Since it is not immediately obvious what in the build directory has to
be removed for a complete rebuild of M³, I've added the clean command
for that purpose now. For completeness, there is also a distclean
command which removes the entire build directory including the cross
compilers.