ninjagen: only add --{start,end}-group if libraries are linked.
Although this is also nicer, because we don't need it if there are no
libraries, there is another reason: the linker on MacOS does not know
these flags and we pass them to the host compiler to compile some tools
as well. Since they currently don't link against any libraries, this
improvement fixes this problem for now.