tools: use bash array to collect gem5 args instead of xargs.
As it turned out, the -o argument for xargs is also not what we want,
because sometimes we don't want to have a tty. For examples, our tests
want to use /dev/null as stdin.

Since we want to control what stdin is and the -a argument for xargs is
not part of any standard, I now removed the usage of xargs altogether.
Instead, we are now building an array in bash and pass the elements as
individual arguments to gem5.