kernel: improved logging of EP configurations.
Previously, we only saw the EP configurations for gate activations.
Sometimes it's also helpful to see the configuration of EPs for
TileMux, syscalls, etc. and therefore, I've reworked the logging now.
With this change, we see all EP configurations on user tiles if the EPS
log is enabled and even those on the kernel tiles if the ALL_EPS log is
enabled.