Reworked the logging approach.
Previously, we defined the logging flags at compile time, scattered over
several files. With this, we were able to completely eliminate disabled
log statements from the code. However, despite the code size and
potentially performance advantages, it was difficult to use
(especially for outsiders) and required lots of recompilations.

As we do no longer optimize for code size since quite some time now, the
approach doesn't make so much sense anymore. Therefore, I've decided
now to rework it so that we can control the logging flags at boot or
runtime. This increases the code size slightly, but does not seem to
have a measurable performance impact.

I considered to use the famous log crate from crates.io at first, but it
turned out that it does not provide advantages for us. This is because
the log crate only defines 6 log levels, which makes it pretty coarse
grained. This may be okay for many applications, but since we are
working with gem5, every unnecessarily executed log statement increases
the simulation time measurably. We thus want to control exactly what
statements are executed. In theory this would be possible by filtering
by the target, which is passed to the logger by the log crate
(typically the target is the module name). However, this would require
us to do several string comparisons for every single log statement just
to figure out whether it should be executed. Therefore, the performance
implications would be severe even if all log flags were disabled.

Instead of using the log crate, I therefore basically kept my approach
that uses several and individually controllable log flags, but made
them changable at boot/runtime. Fortunately, the new bitflags version
provides a way to turn a string like "A|B|C" into an instance of a
bitflags struct with these bits set. This makes it really easy for us,
because no manual parsing is required. We therefore now have a large
LogFlags struct that contains all flags and simply use
LogFlags::from_str to instantiate it. For simplicity and consistency,
the user uses a comma-separated list though.

As logging works for both Rust and C++ we unfortunately need to do some
manual work in C++. It's not that bad though, because pretty much all
logging is done in Rust.

Besides the already existing per-component flags, we also have three
general flags: Info, Debug, and Error. These are used by various
components and Info and Error is enabled by default. These flags are
also used in some applications, so that we don't need to add new flags
for applications.

At runtime, the log flags are passed to all components via the
environment variable LOG. This for example also allows it to pass log
flags to applications started in the M³-Shell:

  $ LOG=Info,LibNet ping 127.0.0.1

Additionally, M3_LOG can be specified for b's run commands to pass log
flags to all components in the system. That is, the kernel and all
TileMux instances will receive it directly and the kernel passes it to
root. Since environment variables are by default just copied to childs,
the child activities receive them as well.

Note that I also renamed M3_GEM5_DBG to M3_GEM5_LOG for consistency.